requirements.txt
# requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0
soundfile>=0.12.0
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
supabase>=2.0.0
python-dotenv>=1.0.0
runpod>=1.3.0
boto3>=1.34.0  # Supabase S3 uyğunluğu üçün